Clean up scope expression templates
